Change how renames were done
